chore(veritech): restructuring into cmd & pkg components
