+++ .github/labeler.yml
@@ -69,8 +69,9 @@ A-si-settings:
   - Cargo.lock
   - rust-toolchain
   - rustfmt.toml
-A-si-veritech:
-  - components/si-veritech/**/*
+A-veritech:
+  - cmd/veritech/**/*
+  - pkg/veritech/**/*
   - .prettierrc.js
 A-si-web:
   - components/si-web-app/**/*
