+++ .run/Veritech.run.xml
@@ -5,7 +5,7 @@
     <option name="SCRIPT_PATH" value="" />
     <option name="SCRIPT_OPTIONS" value="" />
     <option name="INDEPENDENT_SCRIPT_WORKING_DIRECTORY" value="true" />
-    <option name="SCRIPT_WORKING_DIRECTORY" value="$PROJECT_DIR$/components/si-veritech" />
+    <option name="SCRIPT_WORKING_DIRECTORY" value="$PROJECT_DIR$/cmd/veritech" />
     <option name="INDEPENDENT_INTERPRETER_PATH" value="true" />
     <option name="INTERPRETER_PATH" value="/usr/bin/zsh" />
     <option name="INTERPRETER_OPTIONS" value="" />
