+++ cmd/veritech/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "veritech-cli"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.56"
+
+[[bin]]
+name = "veritech"
+path = "src/main.rs"
+
+[dependencies]
+clap = { version = "3.0.0-beta.4" }
+color-eyre = { version = "0.5.11" }
+# TODO(fnichol): I *think* we can boil this off?
+si-settings = { path = "../../components/si-settings" }
+tokio = { version = "1.12.0", features = ["full"] }
+veritech = { path = "../../pkg/veritech", default-features = false, features = ["server"] }
+
+# TODO(fnichol): extract into `pkg/telemetry`
+opentelemetry = { version = "0.16.0", features = ["rt-tokio", "trace"] }
+opentelemetry-otlp = { version = "0.9.0" }
+opentelemetry-semantic-conventions = { version = "0.8.0" }
+thiserror = { version = "1.0" }
+tracing = { version = "0.1" }
+tracing-opentelemetry = { version = "0.15.0" }
+tracing-subscriber = { version = "0.2.19" }
