+++ cmd/veritech/src/telemetry.rs
