+++ components/si-model/src/resolver.rs
@@ -78,7 +78,7 @@ pub enum ResolverError {
     )]
     MismatchedFunctionResultAndSchema(Option<Prop>, serde_json::Value),
     #[error("veritech error: {0:?}")]
-    VeritechError(#[from] si_veritech::client::VeritechClientError),
+    VeritechError(#[from] veritech::client::VeritechClientError),
     #[error("function error: {0} {1}")]
     FunctionError(String, String),
 }
@@ -384,17 +384,16 @@ impl ResolverBinding {
             ResolverBackendKindBinding::Js(context) => {
                 let conn: NatsConn = nats.connection.clone().into();
                 let result =
-                    si_veritech::client::run_function(&conn, "resolver", context.code.clone())
-                        .await?;
+                    veritech::client::run_function(&conn, "resolver", context.code.clone()).await?;
                 match result {
-                    si_veritech::FunctionResult::Success(success) => {
+                    veritech::FunctionResult::Success(success) => {
                         if success.unset {
                             return Ok(None);
                         } else {
                             success.data
                         }
                     }
-                    si_veritech::FunctionResult::Failure(failure) => {
+                    veritech::FunctionResult::Failure(failure) => {
                         return Err(ResolverError::FunctionError(
                             failure.error.name,
                             failure.error.message,
