+++ pkg/veritech/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "si-veritech"
+name = "veritech"
 version = "0.1.0"
 edition = "2018"
 
@@ -10,15 +10,8 @@ default = ["client", "server"]
 client = []
 
 server = [
-  "clap",
-  "color-eyre",
   "derive_builder",
   "futures-lite",
-  "opentelemetry",
-  "opentelemetry-otlp",
-  "opentelemetry-semantic-conventions",
-  "tracing-opentelemetry",
-  "tracing-subscriber",
 ]
 
 [dependencies]
@@ -27,8 +20,9 @@ futures = { version = "0.3.17"}
 serde = "1.0.123"
 serde_json = { version = "1.0.64", features = ["preserve_order"] }
 cyclone = { path = "../../pkg/cyclone", default-features = false, features = ["client"] }
-si-data = { path = "../si-data" }
-si-settings = { path = "../si-settings" }
+si-data = { path = "../../components/si-data" }
+# TODO(fnichol): I *think* we can boil this off?
+si-settings = { path = "../../components/si-settings" }
 thiserror = { version = "1.0" }
 tokio = { version = "1.2.0", features = ["full"] }
 toml = "0.5"
@@ -37,12 +31,5 @@ tracing = { version = "0.1" }
 # client only
 
 # server only
-clap = { version = "3.0.0-beta.4", optional = true }
-color-eyre = { version = "0.5.11", optional = true }
 derive_builder = { version = "0.10.2", optional = true }
 futures-lite = { version = "1.12.0", optional = true }
-opentelemetry = { version = "0.16.0", features = ["rt-tokio", "trace"], optional = true }
-opentelemetry-otlp = { version = "0.9.0", optional = true }
-opentelemetry-semantic-conventions = { version = "0.8.0", optional = true }
-tracing-opentelemetry = { version = "0.15.0", optional = true }
-tracing-subscriber = { version = "0.2.19", optional = true }
