+++ pkg/veritech/DESIGN.md
