+++ pkg/veritech/src/client.rs
