+++ pkg/veritech/src/server/publisher.rs
